// src/types/cliente.ts
export interface ClienteDB {
  id?: number;
  nombre: string;
  tipoDocumento?: string;
  tipo_documento?: string;
  numeroDocumento?: string;
  numero_documento?: string;
  telefono?: string;
  correo?: string;
  direccion?: string;
  estado?: string;
  fechaCreacion?: string;
  fecha_creacion?: string;
}